+++ packages/shared/src/components/profile/UpvoterList.tsx
@@ -1,10 +1,10 @@
 import React, { ReactElement } from 'react';
 import Link from 'next/link';
 import { UseInfiniteQueryResult } from 'react-query';
-import useFeedInfiniteScroll from '../../hooks/feed/useFeedInfiniteScroll';
 import { UpvotesData } from '../../graphql/common';
 import { UpvoterListPlaceholder } from './UpvoterListPlaceholder';
 import { UserShortInfo } from './UserShortInfo';
+import InfiniteScrolling from '../containers/InfiniteScrolling';
 
 export interface UpvoterListProps {
   queryResult: UseInfiniteQueryResult<UpvotesData>;
@@ -16,19 +16,11 @@ export function UpvoterList({
   queryResult,
   ...props
 }: UpvoterListProps): ReactElement {
-  const canFetchMore =
-    !queryResult.isLoading &&
-    !queryResult.isFetchingNextPage &&
-    queryResult.hasNextPage &&
-    queryResult.data.pages.length > 0;
-
-  const infiniteScrollRef = useFeedInfiniteScroll({
-    fetchPage: queryResult.fetchNextPage,
-    canFetchMore,
-  });
-
   return (
-    <div className="flex relative flex-col">
+    <InfiniteScrolling
+      queryResult={queryResult}
+      placeholder={<UpvoterListPlaceholder placeholderAmount={1} />}
+    >
       {queryResult.data.pages.map(
         (page) =>
           page &&
@@ -43,14 +35,7 @@ export function UpvoterList({
             </Link>
           )),
       )}
-      {queryResult.isFetchingNextPage && (
-        <UpvoterListPlaceholder placeholderAmount={1} />
-      )}
-      <div
-        className="absolute bottom-0 left-0 w-px h-px opacity-0 pointer-events-none"
-        ref={infiniteScrollRef}
-      />
-    </div>
+    </InfiniteScrolling>
   );
 }
